Test 2 green.
We make the code also write to col0 pin, still not caring about the clean up.
Note that adding this breaks the easy test one, we need to add that we expect other
pins to be written and ignore them

We add the matchers _ and Not, and also Times(AnyNumber()).